fixed around wrap reviews slider
