+++ src/components/parallax-fest.pug
@@ -1,4 +1,4 @@
-.parallax.parallax-fest
+.parallax-fest
   .parallax-fest__layer(data-speed="0")
     +image('content/parallax/fest/1.png', 'parallax-fest__img', 'Изображение параллакса')
   .parallax-fest__layer(data-speed="0")
