+++ src/components/parallax-music.pug
@@ -1,4 +1,4 @@
-.parallax.parallax-music
+.parallax-music
   .parallax-music__layer(data-speed="0")
     +image('content/parallax/music/1.png', 'parallax-music__img', 'Изображение параллакса')
   .parallax-music__layer(data-speed="0")
