+++ src/main.js
@@ -3,7 +3,6 @@ if (process.env.NODE_ENV === "development") {
   require("file-loader!./index.pug");
 }
 
-import "./scripts/parallax-fest";
 import anchorScroll from  "./scripts/anchorScroll";
 import "./scripts/header";
 import "./scripts/popup-menu";
@@ -11,6 +10,9 @@ import "./scripts/skills";
 import "./scripts/works";
 import "./scripts/reviews";
 import "./scripts/order";
-import "./scripts/paralax-music";
+import parallax from "./scripts/parallax";
 
 anchorScroll();
+
+parallax('.parallax-fest', 'up', 10);
+parallax('.parallax-music', 'down', 100);
