+++ src/styles/blocks/rewievs.pcss
@@ -75,8 +75,13 @@
   }
 }
 
+.reviews__carusel-wrapper {
+  height: 100%;
+}
+
 .reviews__carusel {
   width: 100%;
+  height: 100%;
   max-width: 1000px;
   position: relative;
 }
