+++ src/styles/blocks/works.pcss
@@ -91,6 +91,7 @@
 
 .works__slider_thumbs-item {
   min-width: 33.33%;
+  height: 100%;
   position: relative;
   
   &::after {
@@ -118,11 +119,6 @@
   &.active {
     border: 3px solid $hover-color;
   }
-}
-
-.works__slider_thumbs-btn {
-  width: 100%;
-  height: 100%;
 }
 
 .works__slider_thumbs-pic {
